fix(audit): resolve 4-agent audit findings — concurrency, safety, DRY
Concurrency (C1): Arc<RwLock<Option<Arc<LanceStore>>>> — require_store()
now returns cloned Arc, releasing RwLock guard immediately. Prevents
write-lock starvation when handlers hold read locks across .await.

DRY (C1): Extracted shared parse_evidence_from_record + extract_field
into scoring/evidence.rs. Removed duplicates from server.rs and decay.rs.

Safety (C1): HTTP client timeout 120s. Error messages truncated to 200
chars to prevent API key leakage. DB errors logged via tracing::warn
instead of silently swallowed via unwrap_or_default.

Config: Added temperature field to LlmConfig (default 0.7).
Cleanup: Removed duplicate Request types from types.rs (params are
inline in server.rs). Zero warnings.

168 tests pass, 0 warnings.

Refs: 4-agent audit findings
