+++ crates/forgeplan-core/src/scoring/mod.rs
@@ -1,2 +1,3 @@
 pub mod decay;
+pub mod evidence;
 pub mod reff;
